requirements.txt
python-telegram-bot
httpx
aiomysql
pydub
SpeechRecognition
pytest
pytest-asyncio